[fix]  정렬기준 default 값 변경

+++ src/main/java/org/gachon/checkmate/domain/checkList/entity/CleanType.java
@@ -18,7 +18,7 @@ public enum CleanType implements EnumField {
     private final String desc;
     private final int size = 5;
 
-    public int compareRateTo(CleanType e) {
-        return 1 - Math.abs(Integer.parseInt(this.getCode()) - Integer.parseInt(e.getCode())) / (this.size - 1);
+    public double compareRateTo(CleanType e) {
+        return 1 - (double) Math.abs(Integer.parseInt(this.getCode()) - Integer.parseInt(e.getCode())) / (this.size - 1);
     }
 }
